reduced batch size to 1

+++ scripts/functions/pytorchtools.py
@@ -120,7 +120,7 @@ def collate_voxels(batch):
         #target = target + (0.1**0.5)*torch.randn(target.shape)
 
 
-    yy_pad = torch.stack(yy).to(device)
+    yy_pad = torch.stack(yy).float().to(device)
     xx_pad = torch.stack(xx_pad).to(device)
 
     return xx_pad, yy_pad, x_lens, y_lens
